+++ ChangeLog.md
@@ -11,6 +11,9 @@
   - Rename CConnecterDesktop to CConnecterConnect
   - Add CParameterUI
   - Rename CConnecterDesktopThread to CConnecterThread
+-  FreeRDP: prepare default settings    
+  - Fix ripple [#27](https://github.com/KangLin/RabbitRemoteControl/issues/27) [#21](https://github.com/KangLin/RabbitRemoteControl/issues/27)  
+    Because of the default is not set FreeRDP_SupportGraphicsPipeline feature
 
 ### v0.0.27
 - Client: